Read in company financial data processed from eikon exports and AR master data that contain information on multiple credit risk inputs

Read in company financial data processed from eikon exports and AR master data that contain information on multiple credit risk inputs

Usage

read_financial_data(path = NULL)

Arguments

path

A string that points to the location of the file containing the company financial data.

See also